requirements.txt
crewai
requests
pyyaml
pandas
python-dateutil
sqlalchemy
slack_sdk
twilio



